% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaries.R
\name{closure_count}
\alias{closure_count}
\title{Summarize CLOSURE data}
\usage{
closure_count(data)
}
\arguments{
\item{data}{CLOSURE data frame.}
}
\value{
Tibble (data frame) with these columns:
\itemize{
\item \code{value}: integer. Scale values.
\item \code{n_absolute}: Absolute frequencies of the scale values in \code{data}.
\item \code{n_relative}: Relative frequencies of the scale values in \code{data}.
}
}
\description{
Use \code{closure_count()} to count frequencies of data coming from
\code{closure_read()}.

Should also support data coming from \code{\link[=closure_pivot_longer]{closure_pivot_longer()}} in the
future.
}
