% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count.R
\name{closure_count_all}
\alias{closure_count_all}
\alias{closure_count_initial}
\title{Count CLOSURE samples in advance}
\usage{
closure_count_all(
  mean,
  sd,
  n,
  scale_min,
  scale_max,
  rounding = "up_or_down",
  threshold = 5
)

closure_count_initial(scale_min, scale_max)
}
\arguments{
\item{mean}{String (length 1). Reported mean.}

\item{sd}{String (length 1). Reported sample standard deviation.}

\item{n}{Numeric (length 1). Reported sample size.}

\item{scale_min, scale_max}{Integers (length 1 each). Minimum and maximum of
the scales to which the reported statistics refer.}

\item{rounding}{String (length 1). Rounding method assumed to have created
\code{mean} and \code{sd}. See \href{https://lhdjung.github.io/roundwork/articles/rounding-options.html}{\emph{Rounding options}},
but also the \emph{Rounding limitations} section below. Default is
\code{"up_or_down"} which, e.g., unrounds \code{0.12} to \code{0.115} as a lower bound and
\code{0.125} as an upper bound.}

\item{threshold}{Numeric (length 1). Number from which to round up or down,
if \code{rounding} is any of \code{"up_or_down"}, \code{"up"}, and \code{"down"}. Default is
\code{5}.}
}
\value{
Integer (length 1).
}
\description{
Determine how many samples \code{\link[=closure_generate]{closure_generate()}} would find for
a given set of summary statistics.
\itemize{
\item \code{closure_count_all()} counts all CLOSURE samples that correspond to the
input summary statistics, but without actually generating the samples. This
is much faster than \code{closure_generate()} if there are many results.
\item \code{closure_count_initial()} only counts the first round of samples, from
which all other ones would be generated. Based on scale range only.
}

This can help predict how much time \code{\link[=closure_generate]{closure_generate()}} would take, and
avoid prohibitively long runs.
}
\examples{
closure_count_all(
  mean = "3.5",
  sd = "1.7",
  n = 70,
  scale_min = 1,
  scale_max = 5
)

closure_count_initial(scale_min = 1, scale_max = 5)
}
