% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count.R
\name{closure_count_combinations_initial}
\alias{closure_count_combinations_initial}
\title{Count initial CLOSURE combinations}
\usage{
closure_count_combinations_initial(min_scale, max_scale)
}
\arguments{
\item{min_scale, max_scale}{Integers (length 1 each). Minimum and maximum of
the scales to which the reported statistics refer.}
}
\description{
Count initial CLOSURE combinations
}
