% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count.R
\name{closure_count_initial}
\alias{closure_count_initial}
\title{Count CLOSURE combinations in advance}
\usage{
closure_count_initial(scale_min, scale_max)
}
\arguments{
\item{scale_min, scale_max}{Integers (length 1 each). Minimum and maximum of
the scales to which the reported statistics refer.}
}
\value{
Integer (length 1).
}
\description{
Determine how many combinations \code{\link[=closure_combine]{closure_combine()}} would find
for a given set of summary statistics.
\itemize{
\item TODO: add a \code{closure_count_all()} function.
\item \code{closure_count_initial()} only counts the first round of combinations,
from which all other ones would be generated.
}

This can help predict how much time \code{\link[=closure_combine]{closure_combine()}} would take, and
avoid prohibitively long runs.
}
