% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/horns-analyze.R
\name{closure_horns_analyze}
\alias{closure_horns_analyze}
\title{Horns index for each CLOSURE sample}
\usage{
closure_horns_analyze(data)
}
\arguments{
\item{data}{List returned by \code{\link[=closure_generate]{closure_generate()}}.}
}
\value{
Named list of five tibbles (data frames):
\itemize{
\item \strong{closure_generate_inputs}: Adopted from the input.
\item \strong{horns_metrics}: Summary statistics of the distribution of horns index
values:
\itemize{
\item \code{mean}, \code{uniform}: arithmetic mean.
\item \code{uniform}: horns value of a hypothetical uniform distribution. Equal to
\code{horns_uniform} from the results of \code{closure_generate()}.
\item \code{sd}: double. Standard deviation.
\item \code{cv}: double. Coefficient of variation, i.e., \code{sd / mean}.
\item \code{mad}: double. Median absolute deviation; see \code{\link[stats:mad]{stats::mad()}} but also
details here.
\item \code{min}, \code{median}, \code{max}: double. Minimum, median, and maximum horns
index.
\item \code{range}: double. Equal to \code{max - min}.
}
\item \strong{frequency_horns_min}, \strong{frequency_horns_max}: Frequency tables of the
samples with the minimum or maximum horns values. They are structured like
the \code{frequency} table in the output of \code{closure_generate()}. However, note
that the \code{f_absolute} column here can include values from multiple samples,
so it is better to focus on \code{f_average} or \code{f_relative}. These two refer to
the average sample from the subset of samples with the minimum or maximum
horns value.
\item \strong{horns_results}:
\itemize{
\item \code{id}: integer. Uniquely identifies each horns index, just like their
corresponding samples in \code{closure_generate()}.
\item \code{horns}: double. Horns index for each individual sample.
}
}
}
\description{
Following up on \code{\link[=closure_generate]{closure_generate()}}, you can call
\code{closure_horns_analyze()} to compute the horns index for each individual
sample and calculate summary statistics on the distribution of these
indices. See \code{\link[=horns]{horns()}} for the metric itself.

See \link{horns_plot} for follow-up visualizations.
}
\details{
The results of \code{closure_generate()} already include a \code{"horns"}
metric that is computed from the frequencies of scale values in the mean
sample. This is different from the \code{"mean"} metric found here, which is the
mean of the horns indices of all samples. These two values can differ
because the horns index is based on squared deviations, which can result in
non-linearities.

The \code{"mad"} column is created by \code{\link[stats:mad]{stats::mad()}}, but it does not rely on
this function's default of adjusting the result via multiplication by a
constant (about 1.48). This assumes a normal distribution, which is not
generally the case with horns index values. Here, the constant is set to
\code{1}.
}
\examples{
data <- closure_generate(
  mean = "2.9",
  sd = "0.5",
  n = 70,
  scale_min = 1,
  scale_max = 5
)

closure_horns_analyze(data)
}
