% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaries.R
\name{closure_summarize}
\alias{closure_summarize}
\title{Summarize CLOSURE data}
\usage{
closure_summarize(data)
}
\arguments{
\item{data}{CLOSURE data frame.}
}
\value{
Tibble (data frame) with these columns:
\itemize{
\item \code{value}: integer. Scale values.
\item \code{f_absolute}: Absolute frequencies of the scale values in \code{data}.
\item \code{f_relative}: Relative frequencies of the scale values in \code{data}.
}
}
\description{
Use \code{closure_summarize()} to compute frequencies of data coming
from \code{\link[=closure_combine]{closure_combine()}} or \code{\link[=closure_pivot_longer]{closure_pivot_longer()}}.
}
