% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-write.R
\name{closure_write}
\alias{closure_write}
\alias{closure_read}
\title{Write CLOSURE results to disk (and read them back in)}
\usage{
closure_write(data, path)

closure_read(
  path,
  include = c("stats_only", "stats_and_horns", "capped_error", "all"),
  samples_cap = NULL
)
}
\arguments{
\item{data}{List returned by \code{closure_generate()}.}

\item{path}{String (length 1). File path where \code{closure_write()} will create a
new folder with the results. Set it to \code{"."} to choose the current
working directory. For \code{closure_read()}, the path to an existing
folder with results.}

\item{include}{String (length 1). Which parts of the detailed results should be read in?
\itemize{
\item With \code{"stats_only"}, the default, no results are read.
\item \code{"stats_and_horns"} reads the horns index values, but not the samples.
\item \code{"capped_error"} checks whether the number of samples is higher than a
given threshold (see \code{samples_cap}). If so, it throws an error;
but if not, it reads both the samples and the horns values.
\item \code{"all"} reads both the samples and the horns values.
}}

\item{samples_cap}{Numeric (length 1). When using \code{include = "capped_error"},
enter a whole number here to specify a cap. Default is \code{NULL}.}
}
\value{
\itemize{
\item \code{closure_write()} returns the path to the new folder it created.
\item \code{closure_read()} returns a list of the same kind as
\code{\link[=closure_generate]{closure_generate()}}.
}
}
\description{
You can use \code{closure_write()} to save the results of
\code{\link[=closure_generate]{closure_generate()}} on your computer. A message will show the exact
location.

The data are saved in a new folder as separate Parquet files, one for
each tibble in \code{closure_generate()}'s output. \code{results} is saved as
counts.parquet, along with scale_values.parquet which maps the count
columns to scale values.

\code{closure_read()} is the opposite: it reads those files back into R,
recreating the original CLOSURE list. This is useful for later
analyses if you don't want to re-run a lengthy \code{closure_generate()}
call. It also works with results that \code{closure_generate()} wrote
to disk itself using \code{path = "your/path"}.
}
\details{
\code{closure_write()} saves all tibbles as Parquet files. This is much
faster and takes up far less disk space --- roughly 1\% of a CSV file
with the same data. Speed and disk space can be relevant with large
result sets.

Use \code{closure_read()} to import the CLOSURE list from the folder
back into R. This is based on \code{\link[nanoparquet:read_parquet]{nanoparquet::read_parquet()}}.
}
\section{Folder name}{
 The new folder's name will contain all the inputs that
determine the CLOSURE results. Dashes separate values and
underscores replace decimal periods. For example:

\preformatted{

 CLOSURE-3_5-1_0-90-1-5-up_or_down-5
 }

The order is the same as in \code{closure_generate()}:

\preformatted{

 closure_generate(
   mean = "3.5",
   sd = "1.0",
   n = 90,
   scale_min = 1,
   scale_max = 5,
   rounding = "up_or_down",  # default
   threshold = 5             # default
 )

}
}

\examples{
data <- closure_generate(
  mean = "2.7",
  sd = "0.6",
  n = 45,
  scale_min = 1,
  scale_max = 5
)

# Writing to a temporary folder just for this example.
# You should write to a real folder instead.
# A simple way is path = "." for your current directory.
path_new_folder <- closure_write(data, path = tempdir())

# In a later session, conveniently read the files
# back into R. This returns the original list,
# identical except for floating-point error.
# (Of course, the `path_new_folder` variable will
# no longer be available -- instead, paste the path
# to your folder here.)
closure_read(path_new_folder)
}
