% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{create_combinations}
\alias{create_combinations}
\title{Core CLOSURE implementation, processed for R}
\usage{
create_combinations(
  mean,
  sd,
  n,
  scale_min,
  scale_max,
  rounding_error_mean,
  rounding_error_sd
)
}
\description{
Core CLOSURE implementation, processed for R
}
