% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo-plot.R
\name{demo_plot_bar}
\alias{demo_plot_bar}
\title{Visualize example distributions and their \eqn{h} values}
\usage{
demo_plot_bar(
  freqs,
  format = c("percent", "absolute_percent", "absolute", "relative"),
  samples = c("mean", "all"),
  bar_alpha = 0.75,
  bar_color = "darkgreen",
  show_text = TRUE,
  text_color = bar_color,
  text_size = 12,
  text_offset = 0.05,
  mark_thousand = ",",
  mark_decimal = "."
)
}
\arguments{
\item{freqs}{Numeric. Vector of relative or absolute frequencies to
visualize.}

\item{format}{String (length 1). What should the bars show? The default is
\code{"percent"}. Similarly, \code{"absolute_percent"} shows the count of each scale
value and its percentage of all values. Other options are \code{"absolute"} and
\code{"relative"} frequencies.}

\item{samples}{String (length 1). How to aggregate the samples? Either take
the average sample (\code{"mean"}, the default) or the sum of all samples
(\code{"all"}). This only matters if absolute frequencies are shown.}

\item{bar_alpha}{Numeric (length 1). Opacity of the bars. Default is \code{0.75}.}

\item{bar_color}{String (length 1). Color of the bars. Default is
\code{"darkgreen"}.}

\item{show_text}{Logical (length 1). Should the bars be labeled with the
corresponding frequencies? Default is \code{TRUE}.}

\item{text_color}{String (length 1). Color of the frequency labels. By
default, the same as \code{bar_color}.}

\item{text_size}{Numeric (length 1). Base font size in pt. Default is \code{12}.}

\item{text_offset}{Numeric (length 1). Distance between the text labels and
the bars. Default is \code{0.05}.}

\item{mark_thousand, mark_decimal}{Strings (length 1 each). Delimiters between
groups of digits in text labels. Defaults are \code{","} for \code{mark_thousand}
(e.g., \code{"20,000"}) and \code{"."} for \code{mark_decimal} (e.g., \code{"0.15"}).}
}
\value{
A ggplot object.
}
\description{
The \verb{demo_plot_*()} functions are variants of \verb{closure_plot_*()}
that directly visualize a given frequency distribution. Their purpose is to
illustrate general points about CLOSURE-type techniques and the horns
index, which is also shown; see \code{\link[=horns]{horns()}}.
\itemize{
\item \code{demo_plot_bar()} is like \code{\link[=closure_plot_bar]{closure_plot_bar()}} except it is not faceted.
\item \code{demo_plot_horns_histogram()} is like \code{\link[=closure_plot_horns_histogram]{closure_plot_horns_histogram()}}.
\item \code{demo_plot_ecdf()} is like \code{\link[=closure_plot_ecdf]{closure_plot_ecdf()}}.
}
}
\details{
In keeping with the forensic metascience tradition of tortured
backronyms, DEMO stands for "displaying examples of meticulous operation".
}
\examples{
# Zero variance: h = 0
demo_plot_bar(freqs = c(0, 40, 0, 0, 0))

# Perfect "horns of no confidence": h = 1
demo_plot_bar(freqs = c(20, 0, 0, 0, 20))

# Grouped around h = ~0.44, the uniform horns index
# for 7-point scales
demo_plot_bar(freqs = 21:27)
}
