% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/horns-plot.R
\name{horns_plot}
\alias{horns_plot}
\alias{closure_plot_bar_min_max}
\title{Visualize horns values distribution}
\usage{
closure_plot_bar_min_max(
  data,
  min_max = c("both", "min", "max"),
  frequency = c("absolute-percent", "absolute", "relative", "percent"),
  samples = c("mean", "all"),
  facet_labels = c("Minimal variability", "Maximal variability"),
  facet_labels_parens = "h",
  bar_alpha = 0.75,
  bar_color = "#5D3FD3",
  show_text = TRUE,
  text_color = bar_color,
  text_size = 12,
  text_offset = 0.05,
  mark_thousand = ",",
  mark_decimal = "."
)
}
\arguments{
\item{data}{List returned by \code{\link[=closure_generate]{closure_generate()}}.}

\item{min_max}{String (length 1). Only in \code{closure_plot_bar_min_max()}. Which
plot(s) to show? Options are \code{"both"} (the default), \code{"min"}, and \code{"max"}.}

\item{frequency}{String (length 1). What should the bars display? The
default, \code{"absolute-percent"}, displays the count of each scale value and
its percentage of all values. Other options are \code{"absolute"}, \code{"relative"},
and \code{"percent"}.}

\item{samples}{String (length 1). How to aggregate the samples? Either take
the average sample (\code{"mean"}, the default) or the sum of all samples
(\code{"all"}). This only matters if absolute frequencies are shown.}

\item{facet_labels}{String (length 2). Only in \code{closure_plot_bar_min_max()}.
Labels of the two individual plots. Set it to \code{NULL} to remove the labels.
Default is \code{c("Minimal variability", "Maximal variability")}.}

\item{facet_labels_parens}{String (length 1). Only in
\code{closure_plot_bar_min_max()}. Italicized part of the facet labels inside
the parentheses. Set it to \code{NULL} to remove the parentheses altogether. See
details. Default is \code{"h"}.}

\item{bar_alpha}{Numeric (length 1). Opacity of the bars. Default is \code{0.75}.}

\item{bar_color}{String (length 1). Color of the bars. Default is
\code{"#5D3FD3"}, a purple color.}

\item{show_text}{Logical (length 1). Should the bars be labeled with the
corresponding frequencies? Default is \code{TRUE}.}

\item{text_color}{String (length 1). Color of the frequency labels. By
default, the same as \code{bar_color}.}

\item{text_size}{Numeric (length 1). Base font size in pt. Default is \code{12}.}

\item{text_offset}{Numeric (length 1). Distance between the text labels and
the bars. Default is \code{0.05}.}

\item{mark_thousand, mark_decimal}{Strings (length 1 each). Delimiters between
groups of digits in text labels. Defaults are \code{","} for \code{mark_thousand}
(e.g., \code{"20,000"}) and \code{"."} for \code{mark_decimal} (e.g., \code{"0.15"}).}

\item{binwidth}{Numeric (length 1). Only in
\code{closure_plot_horns_histogram()}. Width of the bins that divide up the
x-axis, passed on to \code{\link[ggplot2:geom_histogram]{ggplot2::geom_histogram()}}. Default is \code{0.01}.}
}
\value{
A ggplot object.
}
\description{
Two plot functions that follow up on \code{\link[=closure_generate]{closure_generate()}}:
\itemize{
\item \code{closure_plot_bar_min_max()} draws barplots of the mean samples from among
those with the minimum or maximum horns index (\eqn{h}). It displays the
typical sample with the least or most amount of variability from among all
CLOSURE samples.
\item \code{closure_plot_horns_histogram()} draws a quick barplot of the distribution
of horns values. The scale is fixed between 0 and 1, so it is aligned with
the range of the horns index. This reveals the big picture, putting any
variability among horns values into perspective.
}
}
\details{
By default, both faceted plots in \code{closure_plot_bar_min_max()} have
a label that includes their horns index (\eqn{h}); see \code{\link[=horns]{horns()}}. You can
remove the parenthesized part using \code{facet_labels_parens = NULL} or the
entire label using \code{facet_labels = NULL}.

Although \code{facet_labels_parens} enables you to choose a different string
inside the parentheses than the default \code{"h"}, this might not be advisable:
if the parentheses are present, they will always display the horns index.
}
\examples{
# Preparation: run CLOSURE
data <- closure_generate(
  mean = "2.9",
  sd = "0.5",
  n = 70,
  scale_min = 1,
  scale_max = 5
)

# Even with minimal and maximal variability,
# the results are almost the same:
closure_plot_bar_min_max(data)

# They cluster in a narrow slice of the 0-1 range
# of the horns index:
closure_plot_horns_histogram(data)
}
