% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{n_to_integer}
\alias{n_to_integer}
\title{Convert strings such as "n1" and "n2" to integers such as 1 and 2}
\usage{
n_to_integer(strings)
}
\description{
Convert strings such as "n1" and "n2" to integers such as 1 and 2
}
