% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine.R
\name{closure_combine}
\alias{closure_combine}
\title{Create CLOSURE combinations}
\usage{
closure_combine(
  mean,
  sd,
  n,
  scale_min,
  scale_max,
  rounding_error_mean,
  rounding_error_sd
)
}
\arguments{
\item{rounding_error_sd}{}
}
\description{
Create CLOSURE combinations
}
\examples{
closure_combine(
  mean = 5.0,
  sd = 2.78,
  n = 30,
  scale_min = 1,
  scale_max = 8,
  rounding_error_mean = 0.01,
  rounding_error_sd = 0.01
)
}
