% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count.R
\name{closure_count_initial}
\alias{closure_count_initial}
\title{Count initial CLOSURE combinations}
\usage{
closure_count_initial(scale_min, scale_max)
}
\arguments{
\item{scale_min, scale_max}{Integers (length 1 each). Minimum and maximum of
the scales to which the reported statistics refer.}
}
\description{
Count initial CLOSURE combinations
}
