% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count.R
\name{closure_count_initial}
\alias{closure_count_initial}
\title{Count CLOSURE samples in advance}
\usage{
closure_count_initial(scale_min, scale_max)
}
\arguments{
\item{scale_min, scale_max}{Integers (length 1 each). Minimum and maximum of
the scales to which the reported statistics refer.}
}
\value{
Integer (length 1).
}
\description{
Determine how many samples \code{\link[=closure_generate]{closure_generate()}} would find for
a given set of summary statistics.
\itemize{
\item \code{closure_count_initial()} only counts the first round of samples, from
which all other ones would be generated.
\item There is currently no \code{closure_count_all()} function.
}

This can help predict how much time \code{\link[=closure_generate]{closure_generate()}} would take, and
avoid prohibitively long runs.
}
\examples{
closure_count_initial(scale_min = 1, scale_max = 5)
}
