% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{closure_plot_bar}
\alias{closure_plot_bar}
\title{Visualize CLOSURE data in a histogram}
\usage{
closure_plot_bar(
  data,
  frequency = c("absolute-percent", "absolute", "relative", "percent"),
  bar_alpha = 0.8,
  bar_color = "royalblue1",
  show_text = TRUE,
  text_offset = 0.05,
  text_color = bar_color
)
}
\arguments{
\item{data}{CLOSURE data frame.}

\item{frequency}{String (length 1). What should the bars display? The
default, \code{"absolute-percent"}, is to show counts of each scale values
together with their percentages of all values. Other options are
\code{"absolute"}, \code{"relative"}, and \code{"percent"}, but all of them display less
information.}

\item{bar_alpha}{Numeric (length 1). Opacity of the bars. Default is \code{0.8}.}

\item{bar_color}{String (length 1). Color of the bars. Default is
\code{"royalblue1"}.}

\item{show_text}{Logical (length 1). Should the bars be labeled with the
corresponding frequencies? Default is \code{TRUE}.}

\item{text_offset}{Numeric (length 1). Distance between the text labels and
the bars. Default is \code{0.05}.}

\item{text_color}{String (length 1). Color of the text labels. By default,
the same as \code{bar_color}.}
}
\value{
A ggplot object.
}
\description{
Call \code{closure_plot_bar()} to get a barplot of data coming from
\code{\link[=closure_combine]{closure_combine()}}, \code{\link[=closure_pivot_longer]{closure_pivot_longer()}}, or
\code{\link[=closure_summarize]{closure_summarize()}}.

For each scale value, the bars show how often this value appears in the
full list of possible raw data combinations found by the CLOSURE algorithm.
}
\examples{
# Create CLOSURE data first:
data <- closure_combine(
  mean = 5.0,
  sd = 2.78,
  n = 30,
  scale_min = 1,
  scale_max = 8,
  rounding_error_mean = 0.01,
  rounding_error_sd = 0.01
)

# Visualize:
closure_plot_bar(data)
}
\seealso{
\itemize{
\item \code{\link[=closure_summarize]{closure_summarize()}}, which displays the same information in a
data frame.
\item \code{\link[=closure_plot_ecdf]{closure_plot_ecdf()}}, an alternative visualization.
}
}
