% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{closure_plot_ecdf}
\alias{closure_plot_ecdf}
\title{Visualize CLOSURE data in an ECDF plot}
\usage{
closure_plot_ecdf(data, line_color = "royalblue1", reference_line_alpha = 0.6)
}
\arguments{
\item{data}{CLOSURE data frame.}

\item{line_color}{String (length 1). Color of the ECDF line. Default is
\code{"royalblue1"}.}

\item{reference_line_alpha}{Numeric (length 1). Opacity of the diagonal
reference line. Default is \code{0.6}.}
}
\value{
A ggplot object.
}
\description{
Call \code{closure_plot_ecdf()} to visualize data coming from
\code{\link[=closure_combine]{closure_combine()}} or \code{\link[=closure_pivot_longer]{closure_pivot_longer()}} using the data's
empirical cumulative distribution function (ECDF).

A diagonal reference line benchmarks the ECDF against a hypothetical linear
relationship.

See \code{\link[=closure_plot_bar]{closure_plot_bar()}} for more intuitive visuals.
}
\details{
Unlike in \code{\link[=closure_plot_bar]{closure_plot_bar()}}, \code{data} cannot currently be output
of \code{\link[=closure_summarize]{closure_summarize()}}.

The present function was inspired by \code{\link[rsprite2:plot_distributions]{rsprite2::plot_distributions()}}.
However, \code{plot_distributions()} shows multiple lines because it is based on
SPRITE, which draws random samples of possible datasets. CLOSURE is
exhaustive, so \code{closure_plot_ecdf()} shows all possible datasets in a
single line.
}
\examples{
# Create CLOSURE data first:
data <- closure_combine(
  mean = 5.0,
  sd = 2.78,
  n = 30,
  scale_min = 1,
  scale_max = 8,
  rounding_error_mean = 0.01,
  rounding_error_sd = 0.01
)

# Visualize:
closure_plot_ecdf(data)
}
