% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{closure_write}
\alias{closure_write}
\title{Write CLOSURE results to disk}
\usage{
closure_write(data, path = ".")
}
\arguments{
\item{data}{List returned by \code{closure_combine()}.}

\item{path}{String (length 1). File path where a new folder with the results
will be created. By default, the current working directory.}
}
\value{
No return value, called for side effects.
}
\description{
You can use \code{closure_write()} to save the results of
\code{\link[=closure_combine]{closure_combine()}} on your computer. A message will show the exact
location.

The data are saved in a new folder as four separate files, one for each
tibble in \code{closure_combine()}'s output. The folder is named after the
parameters of \code{closure_combine()}.
}
\details{
The first three tibbles are saved as CSVs, but the \code{"results"}
tibble becomes a Parquet file. This is much faster and takes up far less
disk space --- roughly 1\% of a CSV file with the same data. Speed and disk
space can be relevant with large result sets.

Opening a Parquet file requires a special reader. Inside of R, you could
use \code{\link[nanoparquet:read_parquet]{nanoparquet::read_parquet()}}.
}
\section{Folder name}{
 The new folder's name should be sufficient to recreate
its CLOSURE results. Dashes separate values, underscores replace decimal
periods. For example:

\preformatted{CLOSURE-3_5-1_0-90-1-5-up_or_down-5}

The order is the same as in \code{closure_combine()}:

\preformatted{
  closure_combine(
    mean = "3.5",
    sd = "1.0",
    n = 90,
    scale_min = 1,
    scale_max = 5,
    rounding = "up_or_down",  # default
    threshold = 5             # default
  )
 }
}

