% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/horns-plot.R
\name{horns-frequency}
\alias{horns-frequency}
\alias{closure_plot_horns_density}
\alias{closure_plot_horns_histogram}
\title{Visualize horns index (\eqn{h}) frequencies}
\usage{
closure_plot_horns_density(
  data,
  alpha = 0.75,
  color = "#5D3FD3",
  reference_line_color = "red",
  text_size = 12
)

closure_plot_horns_histogram(
  data,
  binwidth = 0.01,
  alpha = 0.75,
  color = "#5D3FD3",
  reference_line_color = "red",
  text_size = 12
)
}
\arguments{
\item{data}{List returned by \code{\link[=closure_generate]{closure_generate()}} or \code{\link[=closure_read]{closure_read()}}.}

\item{alpha}{Numeric (length 1). Opacity of the density or bars. Default is
\code{0.75}.}

\item{color}{String (length 1). Color of the density or bars. Default is
\code{"#5D3FD3"}, a purple color.}

\item{text_size}{Numeric (length 1). Base font size in pt. Default is \code{12}.}

\item{binwidth}{Numeric (length 1). Only in \code{closure_plot_horns_histogram()}.
Width of the bins that divide up the x-axis, passed on to
\code{\link[ggplot2:geom_histogram]{ggplot2::geom_histogram()}}. Default is \code{0.01}.}
}
\description{
Two functions that visualize the distribution of horns index
values found by CLOSURE:
\itemize{
\item \code{closure_plot_horns_density()} smooths the distribution, emphasizing its
shape. This is generally recommended.
\item \code{closure_plot_horns_histogram()} bins the distribution into categories,
emphasizing individual groups of values. You can adjust this via the
\code{binwidth} argument.
}

Both mark the minimum and maximum values. The x-axis always ranges from 0
to 1. This reveals the big picture, putting any variability among horns
values into perspective.
}
