% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/horns.R
\name{horns}
\alias{horns}
\alias{horns_uniform}
\alias{horns_rescaled}
\title{Horns index (\eqn{h})}
\usage{
horns(freqs, scale_min, scale_max)

horns_uniform(scale_min, scale_max)

horns_rescaled(freqs, scale_min, scale_max)
}
\arguments{
\item{freqs}{Numeric. Vector with the frequencies (relative or absolute) of
binned observations; e.g., a vector with 5 elements for a 1-5 scale.}

\item{scale_min, scale_max}{Numeric (length 1 each). Minimal and maximal
possible values. For example, with a 1-7 Likert scale, use \code{scale_min = 1}
and \code{scale_max = 7}. Prefer the empirical min and max if available: they
constrain the possible values further.}
}
\value{
Numeric (length 1).
}
\description{
\code{horns()} measures the dispersion in ordinal data based on scale
limits or min/max values. The result is the actual variance as a proportion
of the maximum possible variance. It ranges from 0 to 1:
\itemize{
\item 0 means no variance, i.e., all observations have the same value.
\item 1 means that the observations are evenly split between the extremes, with
none in between.
}

\code{horns_uniform()} computes the value that \code{horns()} would return for a
uniform distribution within given scale limits. This can be useful as a
point of reference for \code{horns()}.

These two functions correspond to the \code{horns} column of \code{results} and the
\code{uniform} column of \code{metrics_horns} in \code{\link[=closure_generate]{closure_generate()}}'s output.

\code{horns_rescaled()} is a version of \code{horns()} that  is normalized by scale
length, such that \code{0.5} always indicates a uniform distribution,
independent of the number of scale points. It is meant to enable comparison
across scales of different lengths, but it is harder to interpret for an
individual scale. This makes it unlikely to be useful in most cases. Even
so, the range and the meaning of \code{0} and \code{1} are the same as for \code{horns()}.
}
\details{
The horns index \eqn{h} is defined as:

\deqn{ h = \frac {\sum_{i=1}^{k} f_i (i - \bar{s})^2} {\frac{1}{4} (k -
1)^2} }

where \eqn{k} is the number of scale points (i.e., the length of \code{freqs}
here), \eqn{f_i} is the relative frequency of the \eqn{i}th point on an
integer scale from \eqn{1} to \eqn{k}, and \eqn{\bar{s}} is the weighted
mean frequency. The mean is derived as follows:

\deqn{ \bar{s} = \sum_{i=1}^{k} i \ f_i }

Note that \eqn{h} only depends on the frequency distribution, not the
actual values of the scale points. This is why both formulas invariably use
a scale from \eqn{1} to \eqn{k}. The only reason why the \code{horns()} function
still takes \code{scale_min} and \code{scale_max} arguments is safety: if \code{freqs} is
misstated such that its length is different from the number of points on
the scale implied by those two arguments, there will be an error.
\subsection{Maximum possible variance}{

The "maximum possible" variance here is based on scale range only
(Popoviciu 1935):

\deqn{ \sigma_{\max}^2 = \frac{1}{4} (k - 1)^2 }

It is deliberately agnostic to the mean and the sample size. In the context
of CLOSURE, these statistics are only relevant for generating possible
samples, which then enter the equation via the numerator; just like the
standard deviation. By contrast, the number of scale points, \eqn{k}, is an
intrinsic property of those samples. It is sensible to assess the variance
of the samples by benchmarking it against the greatest variance that can
occur in any sample with the same \eqn{k}.
}

\subsection{Uniform distribution}{

Although \code{horns_uniform()} is implemented as a wrapper around \code{horns()}
that constructs a perfect uniform distribution internally, an equivalent
closed-form solution can be given as

\deqn{ h_u = \frac{k + 1}{3 (k - 1)} }
}

\subsection{"Horns of no confidence"}{

The term \emph{horns index} was inspired by \href{https://jamesheathers.medium.com/sprite-case-study-3-soup-is-good-albeit-extremely-confusing-food-96ea526c488d}{Heathers (2017)}
which defines the "horns of no confidence" as a reconstructed sample "where
an incorrect, impossible or unlikely value set has all its constituents
stacked into its highest or lowest bins to try meet a ludicrously high SD".
In its purest form, this is a case where \eqn{h = 1}, so \code{horns()} would
return \code{1}. However, note that the implications for the plausibility of any
given set of summary statistics depend on the substantive context of the
data (\href{https://peerj.com/preprints/26968/}{Heathers et al. 2018}).

Heathers wrote about SPRITE, but the logic applies equally to CLOSURE.
Indeed, as CLOSURE is exhaustive but SPRITE is not, it is only CLOSURE that
provides certainty about the set of possible samples and their horns
indices.
}

\subsection{Rust implementation}{

The \code{metrics_horns} tibble that is part of the output of
\code{\link[=closure_generate]{closure_generate()}} is not based on the R functions presented here.
Instead, it relies on efficient Rust implementations of the above formulas.
These Rust functions are part of
\href{https://github.com/lhdjung/closure-core}{closure-core}, which mainly
implements CLOSURE but does not currently export horns functions for users.
}
}
\examples{
# For simplicity, all examples use a 1-5 scale and a total N of 300.

# ---- With all values at the extremes

horns(freqs = c(300, 0, 0, 0, 0), scale_min = 1, scale_max = 5)

horns(c(150, 0, 0, 0, 150), 1, 5)

horns(c(100, 0, 0, 0, 200), 1, 5)


# ---- With some values in between

horns(c(60, 60, 60, 60, 60), 1, 5)

horns(c(200, 50, 30, 20, 0), 1, 5)

horns(c(150, 100, 50, 0, 0), 1, 5)

horns(c(100, 40, 20, 40, 100), 1, 5)
}
\references{
Popoviciu, T. (1935). Sur les équations algébriques ayant toutes
leurs racines réelles. \emph{Mathematica} (Cluj), 9, 129-145.
}
