% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-bar.R
\name{sprite_plot_bar}
\alias{sprite_plot_bar}
\title{Visualize SPRITE data in a barplot}
\usage{
sprite_plot_bar(
  data,
  min_max = c("both", "min", "max"),
  format = c("percent", "absolute_percent", "absolute", "relative"),
  samples = c("mean", "all"),
  facet_labels = c("Minimal variance", "Maximal variance"),
  facet_labels_parens = "h",
  bar_alpha = 0.75,
  bar_color = "royalblue1",
  show_text = TRUE,
  text_color = bar_color,
  text_size = 12,
  text_offset = 0.05,
  mark_thousand = ",",
  mark_decimal = "."
)
}
\arguments{
\item{data}{List returned by \code{\link[=sprite_generate]{sprite_generate()}}.}

\item{min_max}{String (length 1). Which plot panel(s) to show? Options are
\code{"both"} (the default), \code{"min"}, and \code{"max"}.}

\item{format}{String (length 1). What should the bars show? The default is
\code{"percent"}. Similarly, \code{"absolute_percent"} shows the count of each scale
value and its percentage of all values. Other options are \code{"absolute"} and
\code{"relative"} frequencies.}

\item{samples}{String (length 1). How to aggregate the samples? Either take
the average sample (\code{"mean"}, the default) or the sum of all samples
(\code{"all"}). This only matters if absolute frequencies are shown.}

\item{facet_labels}{String (length 2). Labels of the two individual panels.
Set it to \code{NULL} to remove the labels. Default is
\code{c("Minimal variance", "Maximal variance")}.}

\item{facet_labels_parens}{String (length 1). Italicized part of the facet
labels inside the parentheses. Set it to \code{NULL} to remove the parentheses
altogether. See details. Default is \code{"h"}.}

\item{bar_alpha}{Numeric (length 1). Opacity of the bars. Default is \code{0.75}.}

\item{bar_color}{String (length 1). Color of the bars. Default is
\code{"#5D3FD3"}, a purple color.}

\item{show_text}{Logical (length 1). Should the bars be labeled with the
corresponding frequencies? Default is \code{TRUE}.}

\item{text_color}{String (length 1). Color of the frequency labels. By
default, the same as \code{bar_color}.}

\item{text_size}{Numeric (length 1). Base font size in pt. Default is \code{12}.}

\item{text_offset}{Numeric (length 1). Distance between the text labels and
the bars. Default is \code{0.05}.}

\item{mark_thousand, mark_decimal}{Strings (length 1 each). Delimiters between
groups of digits in text labels. Defaults are \code{","} for \code{mark_thousand}
(e.g., \code{"20,000"}) and \code{"."} for \code{mark_decimal} (e.g., \code{"0.15"}).}
}
\value{
A ggplot object.
}
\description{
Call \code{sprite_plot_bar()} to get a barplot of SPRITE results.

For each scale value, the bars show how often this value appears in the
mean samples with the minimum or maximum horns index (\eqn{h}).
This displays the typical sample with the least or most amount
of variance from among all SPRITE samples.
}
\examples{
# Run SPRITE first
data <- sprite_generate(
  mean = "3.0",
  sd = "1.0",
  n = 120,
  scale_min = 1,
  scale_max = 5,
  stop_after = 150
)

sprite_plot_bar(data)
}
